+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+1. `TestScheduler` can `advance` by `TimeInterval`. (#828)
+
 1. Fixed issue with `SignalProducer.Type.interval()` making Swift 5.3 a requirement. (#823 kudos to @mluisbrown) 
 
 *Please add new entries at the top.*